+++ tests/Fakes/Mail.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ShiftOneLabs\LaravelSqsFifoQueue\Tests\Fakes;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use ShiftOneLabs\LaravelSqsFifoQueue\Bus\SqsFifoQueueable;
+
+class Mail extends Mailable implements ShouldQueue
+{
+    use Queueable, SerializesModels, SqsFifoQueueable;
+
+    //
+}